Removed writing to file
